more tests are added and model fixes

+++ RemittanceAPI/V1/Models/Response/FeeResponse.cs
@@ -2,8 +2,8 @@
 {
     public class FeeResponse
     {
-        public int Amount { get; set; }
-        public int Fee { get; set; }
+        public decimal Amount { get; set; }
+        public decimal Fee { get; set; }
     }
 
 }
